refactor: extract diagnostic formatting into shared @utils module
Move countBySeverity, formatCounts, formatGroupedSections, formatMessageList,
and getSeverityLabel from @frontend/vscode/vscodeDiagnostics into a new
platform-agnostic module at @utils/diagnostics/diagnosticFormatting.

This eliminates the duplicated formatting code in leanTypes.ts (-90 lines)
and fixes the @frontend/ import violation in DiagnosticsTool.ts. Both
vscodeDiagnostics.ts and leanTypes.ts now re-export from the shared module.
